initial splitting out of test code
